backend/requirements.txt
# Django Core
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
PyMySQL==1.1.0

# Authentication & Security
argon2-cffi==23.1.0
PyJWT==2.8.0
django-allauth==0.57.0

# Environment Variables
python-dotenv==1.0.0

# Report Generation
reportlab==4.0.7
WeasyPrint==60.1

# Rate Limiting
django-ratelimit==4.1.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Development
ipython==8.18.1
django-debug-toolbar==4.2.0

# Production Server
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
